[zsh] Add rgc alias
This supplements the rgc shell script (which remains as a fallback)
with a zsh function that does the same thing but also sets `$pkd`
to the chosen file.  This variable can be used to pass the filename
to further commands without requiring the user to plan ahead or
re-do the manual picking process.

We also add a `vp` alias, which calls vim on `$pkd`