[nix] Update shell.nix to pin more useful version of nixpkgs
Somehow the version locked as 'nixpkgs' in flake.nix is very out of
date, although I can't see where is pulling that old version in.
This means that running `nix-shell` to bootstrap home-manager
pulls in a version of `nix` that is too outdated to be useful.

The main `nixpkgs` in `flake.nix` is actually locked in `flake.lock`
as `nixpkgs_2`; use that instead.